gift-sender/requirements.txt
aiogram==3.21.0
asyncpg==0.29.0
python-dotenv==1.0.0
sqlalchemy[asyncio]==2.0.25
telethon==1.36.0